refactor(lexer): split `compress` pipeline stage into multiple files (#26437)
Pure refactor. No code changes, just moving code around.

Split `compress` pipeline stage into multiple files.

- `avx2.rs` / `generic.rs` - SIMD and generic impls.
- `common.rs` - Code shared between both impls.
- `mod.rs` - Minimal entry point, chooses which impl to load dependent on target.

Additionally, apply the same kinds of stylistic changes as #26390.
